Refactoring Sample methods
There is a conflict between Sample methods that take a double value and
require the caller to generate the random number and Sample methods that
take a random generator number.

RangedUniform requires multiple random numbers and I anticipate other
classes to do so too, so I can't switch everything to use the passed
value.

The generator methods are more difficult to test. That isn't a great
argument for not using them.

Some distributions -- especially likelihood distributions -- are not
normalized. I don't provide a way to get the total count so a caller
can't generate the selection value correctly anyway.

I think I need to move everything to use internal random generators.

Maybe I can keep the existing ones and add a generator on top. I have to
sum probabilities and do so on every sample. The value methods allow
sample to be called multiple times without recalculating the sum.